fix(quant): min_seconds_between_orders 支持 float（亚秒级 cooldown）
原 type=int 导致 0.5 这种亚秒值无法配置——0 是没冷却（散户挤兑下 backend
可能限流 / 单笔 cap 仍兜底），1 又错失短时多笔成交的机会。

改成 float 后可以填 0.3 / 0.5 等折中值。trader.py 的 int() 转换同步改成
float()。risk.py 的 time.monotonic() - last < cfg 浮点比较本来就工作。

13 risk+broker 测试无回归（原测试 0 / 3 都是合法 float）。

+++ quant/thccb_quant/trader.py
@@ -207,7 +207,7 @@ async def main_async(dry_run: bool) -> int:
             daily_loss_cap_cny=Decimal(str(config["risk"]["daily_loss_cap_cny"])),
             daily_turnover_cap_cny=Decimal(str(config["risk"]["daily_turnover_cap_cny"])),
             max_slippage_bps=int(config["risk"]["max_slippage_bps"]),
-            min_seconds_between_orders=int(config["risk"]["min_seconds_between_orders"]),
+            min_seconds_between_orders=float(config["risk"]["min_seconds_between_orders"]),
         ),
         STATE_DIR,
     )
